requirements.txt
flask
pyembroidery
Pillow
gunicorn
